+++ src/Client/wwwroot/js/bookStore.js
@@ -6,7 +6,9 @@
 
   const db = idb.openDB("Books", 2, {
     upgrade(db) {
-      db.deleteObjectStore(bookStore);
+      if (oldVersion == 1) {
+        db.deleteObjectStore(bookStore);
+      }
 
       db.createObjectStore(bookStore, { keyPath: "id" }).createIndex("lastUpdated", "lastUpdated");
       db.createObjectStore(deleteStore, { keyPath: "deleted" });
